+++ tergite_autocalibration/lib/nodes/characterization/purity_benchmarking/measurement.py
@@ -18,8 +18,8 @@ from quantify_scheduler.operations.gate_library import Measure, Reset, X90, H, X
 from quantify_scheduler.schedules.schedule import Schedule
 
 from tergite_autocalibration.lib.base.measurement import BaseMeasurement
-from tergite_autocalibration.utils.extended_gates import Rxy_12
-from tergite_autocalibration.utils.extended_transmon_element import ExtendedTransmon
+from tergite_autocalibration.utils.dto.extended_gates import Rxy_12
+from tergite_autocalibration.utils.dto.extended_transmon_element import ExtendedTransmon
 import tergite_autocalibration.utils.clifford_elements_decomposition as cliffords