+++ tergite_autocalibration/lib/nodes/qubit_control/spectroscopy/measurement.py
@@ -26,8 +26,8 @@ from quantify_scheduler.resources import ClockResource
 from quantify_scheduler.schedules.schedule import Schedule
 
 from tergite_autocalibration.lib.base.measurement import BaseMeasurement
-from tergite_autocalibration.utils.extended_gates import Measure_RO1
-from tergite_autocalibration.utils.extended_transmon_element import ExtendedTransmon
+from tergite_autocalibration.utils.dto.extended_gates import Measure_RO1
+from tergite_autocalibration.utils.dto.extended_transmon_element import ExtendedTransmon
 
 
 class Two_Tones_Multidim(BaseMeasurement):