+++ tergite_autocalibration/lib/nodes/qubit_control/spectroscopy/node.py
@@ -22,8 +22,7 @@ from tergite_autocalibration.lib.nodes.qubit_control.spectroscopy.measurement im
 )
 from tergite_autocalibration.lib.nodes.schedule_node import ScheduleNode
 
-# TODO: check input
-from tergite_autocalibration.utils.user_input import qubit_samples
+from tergite_autocalibration.lib.utils.samplespace import qubit_samples
 
 
 class Qubit_01_Spectroscopy_Multidim_Node(ScheduleNode):