+++ tergite_autocalibration/lib/nodes/readout/punchout/node.py
@@ -17,7 +17,7 @@ from tergite_autocalibration.lib.nodes.readout.punchout.analysis import (
 )
 from tergite_autocalibration.lib.nodes.readout.punchout.measurement import Punchout
 from tergite_autocalibration.lib.nodes.schedule_node import ScheduleNode
-from tergite_autocalibration.utils.user_input import resonator_samples
+from tergite_autocalibration.lib.utils.samplespace import resonator_samples
 
 
 class Punchout_Node(ScheduleNode):