+++ tergite_autocalibration/lib/nodes/readout/ro_amplitude_optimization/measurement.py
@@ -24,7 +24,7 @@ from quantify_scheduler.operations.pulse_library import (
 from quantify_scheduler.resources import ClockResource
 
 from tergite_autocalibration.lib.base.measurement import BaseMeasurement
-from tergite_autocalibration.utils.extended_transmon_element import ExtendedTransmon
+from tergite_autocalibration.utils.dto.extended_transmon_element import ExtendedTransmon
 
 
 class RO_amplitude_optimization(BaseMeasurement):