+++ tergite_autocalibration/lib/utils/node_factory.py
@@ -18,7 +18,7 @@ from pathlib import Path
 from typing import Dict, List, Union, TYPE_CHECKING
 
 from .reflections import find_inheriting_classes_ast_recursive, import_class_from_file
-from ...utils.regex import camel_to_snake
+from tergite_autocalibration.utils.misc.regex import camel_to_snake
 
 if TYPE_CHECKING:
     from ..base.node import BaseNode