+++ tergite_autocalibration/tools/dataset_browser/quantifiles/plot/colorplot.py
@@ -1,11 +1,11 @@
 from __future__ import annotations
-from typing import List, Sequence
 
-from PyQt5 import QtCore
+from typing import Sequence
+
 import numpy as np
 import pyqtgraph
 import xarray as xr
-
+from PyQt5 import QtCore
 from quantifiles import units
 from quantifiles.plot.baseplot import BasePlot
 from quantifiles.plot.utils import set_label