+++ tergite_autocalibration/utils/hardware_generator.py
@@ -10,9 +10,6 @@
 # copyright notice, and modified files need to carry a notice indicating
 # that they have been altered from the originals.
 
-import csv
-import json
-
 
 class HW_Config_Generator:
     def __init__(
@@ -148,10 +145,10 @@ class HW_Config_Generator:
 
 
 if __name__ == "main":
-    from tergite_autocalibration.config.settings import CONFIG_DIR
+    from tergite_autocalibration.config.globals import ENV
 
-    mixer_file = CONFIG_DIR / "initial.csv"
-    json_config_file = CONFIG_DIR / "HARDWARE_CONFIGURATION_LOKIA_20240504.json"
+    mixer_file = ENV.config_dir / "initial.csv"
+    json_config_file = ENV.config_dir / "HARDWARE_CONFIGURATION_LOKIA_20240504.json"
     CLUSTER_NAME = "clusterA"
     HW_CONFIG = {}
     HW_CONFIG["backend"] = "quantify_scheduler.backends.qblox_backend.hardware_compile"